+++ R/bayes2stage-package.R
@@ -3,11 +3,10 @@
 
 ## usethis namespace: start
 #' @rawNamespace import(data.table, except = c("cube"))
-#' @import nimble
-#' @import nimbleMacros
-#' @import nimbleEcology
 #' @import MASS
-#' @importFrom stats filter qnorm quantile sigma
+#' @importFrom cli cli_h1 cli_h2 cli_alert_info cli_alert_success cli_alert_warning cli_alert_danger
+#' @importFrom posterior as_draws_df
+#' @importFrom stats filter qnorm quantile sigma rnorm rbinom rnbinom rbeta plogis dnorm sd var median shapiro.test
 #' @importFrom purrr map_chr map_lgl
 #' @importFrom instantiate stan_package_model
 ## usethis namespace: end